move read_events and related functions to catalog.py
