+++ obspy/core/event/__init__.py
@@ -34,8 +34,8 @@ from .base import (_event_type_class_factory, Amplitude,
                    QuantityError, ResourceIdentifier,
                    SourceTimeFunction, StationMagnitude,
                    StationMagnitudeContribution, Tensor,
-                   TimeWindow, WaveformStreamID, read_events)
-from .catalog import Catalog
+                   TimeWindow, WaveformStreamID)
+from .catalog import Catalog, read_events
 from .event import Event
 from .radpattern import plot_3drpattern, farfield_p, farfield_s
